refactor: Move get_secret from neurochain utils to shared
