+++ nodes/neurochain/storage/s3/download_from_s3.py
@@ -4,7 +4,8 @@ import boto3
 import requests
 
 from ....categories import S3_CAT
-from ...utils import COMFY_IMAGES_DIR, get_secret
+from ....shared import get_secret
+from ...utils import COMFY_IMAGES_DIR
 
 
 class DownloadFromS3:
